Extract theme-kit utils into a top-level modules
These will become a part of the theme kit package
