+++ .prettierignore
@@ -2,3 +2,4 @@
 build/
 node_modules/
 README.md
+tsconfig.json
