+++ theme-kit/lib/index.ts
@@ -1,3 +1,2 @@
-export { asyncShape } from './asyncShape';
 export { createContentDeliveryClient } from './createContentDeliveryClient';
 export { isUuid } from './isUuid';
