+++ theme-kit/lib/locales.js
