+++ theme/_layout.tsx
@@ -4,7 +4,10 @@ import type { ReactNode } from 'react';
 export async function Layout(props: { children: ReactNode; locale: Culture['code'] }) {
     return (
         <html lang={props.locale}>
-            <body>{props.children}</body>
+            <body>
+                <h1>Root layout</h1>
+                {props.children}
+            </body>
         </html>
     );
 }
