+++ tsconfig.json
@@ -29,6 +29,7 @@
       "@/modules/*": ["modules/*"],
       "@/public/*": ["public/*"],
       "@/theme/*": ["theme/*"],
+      "@/theme-kit": ["theme-kit/"],
       "@/utils": ["utils/"],
       "@/utils/*": ["utils/*"],
       "@/ui": ["ui/"]
