+++ utils/index.ts
@@ -5,4 +5,3 @@ export * from './getTypedKeys';
 export * from './isTrackingEnabled';
 export * from './lang';
 export * from './makeComposableComponent';
-export { api, env, match, route } from './theme-kit';
